move parser configuration into listen and connect methods
The HTTP connection is yielded too late to configure the underlying
parser. It appears that the the first request is parsed prior to the
connection becoming available. The allows configuration values to be
passed in during listen and connect. In the case of listen, the
configuration will be applied to each connection.